refactor(front): move components to feature folder
