+++ front/src/app/logs/[id]/page.tsx
@@ -1,5 +1,5 @@
-import { CardAtom } from "@/atomic/atoms/CardAtom";
 import { LogsData } from "@/data/LogsData";
+import { LogDetailsTable } from "@/features/LogTables/LogDetailsTable";
 
 export default async function LogDetails({
     params,
@@ -8,38 +8,5 @@ export default async function LogDetails({
 }) {
     const logDetails = await LogsData.getLogDetails(params.id);
 
-    return (
-        <CardAtom>
-            <table className="table-auto w-full">
-                <tbody>
-                    <tr>
-                        <td className="font-bold px-4 py-2">ID:</td>
-                        <td className="px-4 py-2">{logDetails.id}</td>
-                    </tr>
-                    <tr>
-                        <td className="font-bold px-4 py-2">
-                            Date/Time (UTC):
-                        </td>
-                        <td className="px-4 py-2">
-                            {new Date(logDetails.datetime_utc).toLocaleString()}
-                        </td>
-                    </tr>
-                    <tr>
-                        <td className="font-bold px-4 py-2">Input String:</td>
-                        <td className="px-4 py-2">{logDetails.input_string}</td>
-                    </tr>
-                    <tr>
-                        <td className="font-bold px-4 py-2">Output String:</td>
-                        <td className="px-4 py-2">
-                            {logDetails.output_string}
-                        </td>
-                    </tr>
-                    <tr>
-                        <td className="font-bold px-4 py-2">Total Tokens:</td>
-                        <td className="px-4 py-2">{logDetails.total_tokens}</td>
-                    </tr>
-                </tbody>
-            </table>
-        </CardAtom>
-    );
+    return <LogDetailsTable logDetails={logDetails} />;
 }
