+++ front/src/features/LogTables/EmptyLogs.tsx
