+++ front/src/features/LogTables/LogsTable.tsx
